Django/mysite/requirements.txt
django
djangorestframework
environs
mysqlclient  # for working with mariadb